Adicionado: menu de ranking, funções de carregar e mostrar ranking. Corrigido: alguns menus

+++ struct_ranking.h
@@ -4,7 +4,7 @@ using namespace std;
 
 struct ranking{
 		string nome;
-		int tempo;
+		float tempo;
 		int modo;
 };
